582 jira to sg poc (#5)
Implemented a proof of concept Jira to Shotgun task syncing to validate the overall code structure. The actual syncing is not yet implemented, except for the Issue creation.
Tweaked Shotgun to Jira routing to support easily webhooks when they will be available.
Added the target issue type as a param for the Task syncer and deprecated the Shotgun custom "jira type" field.
Tweaked the request payload sent by the SG trigger to be compatible with what we will get from webhooks.
Added a mocked Jira to be able to run some tests without any Jira connection.
Added routing tests with a mocked server and request handler.
Improved sync tests.