613 explicit sync (#10)
Changed syncing to be explicit, based on `sg_sync_in_jira` field.

Introduced a `EnableSyncingHandler`, which combines the `TaskIssueHandler` and the `NoteCommentHandler` to perform an initial sync when the checkbox is turned on.
Changed the note syncing logic to take the Task. sg_sync_in_jira value into account.

Moved a couple of methods around to make the `EntityIssueHandler` leaner.
Fixed a couple of bugs.
Fixed the "space in tag names" bug by rejecting them.