Explicit syncing, fixes and improvements (#12)
## Explicit syncing
Task syncing is now explicit and based on a "Sync in Jira" checkbox.

A new `EnableSyncHandler` handler was introduced, which combines the `TaskIssueHandler` and the `NoteCommentHandler` and is triggered when a Task "Sync in Jira" is turned on. A full sync then happens. All Notes linked to the Task are automatically synced. Users don't have to manually select the Notes to sync in Jira.  An alternative to this would be to have a similar checkbox on Notes as well.

Intended workflow:
- Users create a project in Jira.
- Users set the Shotgun project sg_jira_key with the Jira project key.
- They manually decide which Tasks should be synced by clicking on "Sync in Jira". Everything linked to the Task (notes, worklog) is automatically synced as well.

## Fixes
Tags with spaces in their name are rejected as spaces are not allowed in Jira labels.
Fixed a couple of bugs which were introduced by the previous refactor.

## Improvements

Made the `EntityIssueHandler` code leaner by moving a couple of methods around. (960 lines, which seems ok).