Jira Comment to Shotgun Note Syncing and deleted stuff (#15)
## Deleted stuff

Turned a couple of `ValueError` into warning logging to keep things consistent with other places where we just do this where we can't find the expected SG entity or Jira Resource, and to not report un-necessary errors.

## Jira Comment to Shotgun Note Syncing
Comment updates on Jira Issues are now synced back to the corresponding Shotgun Note if there is one (the Shotgun to Jira implementation already exists).

Since Jira does not have a separate subject field for comments, Jira Comments that were created by the SG Jira Bridge currently have a specific format to mimic the Subject and Content structure that Shotgun has. 

When syncing back to Shotgun, if the format doesn't match what we expect, and we cannot parse a `subject` and `content` value for the Shotgun Note, we skip the update. We don't attempt to figure out what went wrong in order to avoid creating a mess and potentially syncing incorrect or incomplete values.

No syncing occurs when new Comments are created in Jira nor when Comments are deleted in Jira.

Removed the Shotgun Note url in the synced Jira comment.

## Fixes
Fixed a small issue with the web server returning a 503 error because we weren't sending headers back correctly on POST requests