622 Handle SG Duplicate Entities (#20)
Enforces a new requirement for unique constraints on Shotgun Jira Key fields. This handles the case when duplicating a Shotgun Entity that is synced to Jira would otherwise potentially create multiple Entities with the same Jira Key.

Validates a Jira Issue is linked to the same Shotgun entity that is attempting to sync to it. This helps avoid corrupting and overwriting existing Issue fields if there's some sort of mismatch.