SG-14475: Multithreaded web server in webapp.py (#36)
In order to increase the volume of data that webapp.py can accept before its connection pool runs dry, we've added the Python standard multithreaded mix-in to the bundled webserver. This does not give the server an infinite amount of scalability, but it does increase the threshold quite a bit before Jira and Shotgun are able to overrun its resources.